changed int to big int

+++ migrations/20160416145427_initial.js
@@ -4,7 +4,7 @@ exports.up = function(knex, Promise) {
     table.increments()
     table.string('fullname')
     table.string('email')
-    table.integer('facebookid')
+    table.bigInteger('facebookid')
     table.string('password')
     table.timestamps()
   })
